% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{detroit}
\alias{detroit}
\title{Homicide in Detroit: The role of Firearms.}
\format{A data frame with 13 rows and 14 variables:
\itemize{
  \item FTP: Full-time police per 100,000 population
  \item UEMP: UEMP - \% unemployed in the population
  \item MAN: MAN - number of manufacturing workers in thousands
  \item LIC: LIC - Number of handgun licences per 100,000 population
  \item GR: GR - Number of handgun registrations per 100,000 population
  \item CLEAR: CLEAR - \% homicides cleared by arrests
  \item WM: WM - Number of white males in the population
  \item NMAN: NMAN - Number of non-manufacturing workers in thousands
  \item GOV: GOV - Number of government workers in thousands
  \item HE: HE - Average hourly earnings
  \item WE: WE - Average weekly earnings:
  \item HOM: HOM - Number of homicides per 100,000 of population
  \item ACC: ACC - Death rate in accidents per 100,000 population
  \item ASR: ASR - Number of assaults per 100,000 population
}}
\source{
\url{http://lib.stat.cmu.edu/datasets/detroit}
}
\usage{
detroit
}
\description{
This is the data set called `DETROIT' in the book `Subset selection in
regression' by Alan J. Miller published in the Chapman & Hall series of
monographs on Statistics & Applied Probability, no. 40.   The data are
unusual in that a subset of three predictors can be found which gives a
very much better fit to the data than the subsets found from the Efroymson
stepwise algorithm, or from forward selection or backward elimination.
The original data were given in appendix A of `Regression analysis and its
application: A data-oriented approach' by Gunst & Mason, Statistics
textbooks and monographs no. 24, Marcel Dekker.   It has caused problems
because some copies of the Gunst & Mason book do not contain all of the data,
and because Miller does not say which variables he used as predictors and
which is the dependent variable.   (HOM was the dependent variable, and the
predictors were FTP ... WE)
}
\keyword{datasets}
